Create Random String Match Generator
This program takes a string as input and then will iteratively create strings of the same length as the input string by randomly generating each letter. This will loop until the randomly generated string is the same as the users string, and then will print out how many loops were iterated until the strings matched. The purpose of this program is to show how the number of tries it takes becomes exponentially greater for each additional letter.